requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
#psycopg2-binary==2.9.9
jinja2==3.1.4
python-multipart==0.0.9
SQLAlchemy==2.0.35
alembic==1.13.2
asyncpg==0.29.0
async-timeout==4.0.3
pydantic-settings==2.5.2
pytest==8.3.3
factory_boy==3.3.1
httpx==0.27.2
pytest-asyncio==0.24.0
aiosqlite==0.20.0
asyncpg